Replace VechileType class in external file
